pyproject.toml
[tool.poetry]
name = "chesss"
version = "0.1.0"
description = ""
authors = ["Abdulmalik Maxudov <maxudov.ar@getflow.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
amqp = "5.2.0"
asgiref = "3.8.1"
billiard = "4.2.1"
celery = "5.4.0"
chess = "1.11.1"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
django = "5.1.2"
djangorestframework = "3.15.2"
kombu = "5.4.2"
pika = "1.3.2"
pika-stubs = "0.1.3"
prompt-toolkit = "3.0.48"
python-dateutil = "2.9.0.post0"
six = "1.16.0"
sqlparse = "0.5.1"
stockfish = "3.28.0"
tzdata = "2024.2"
vine = "5.1.0"
wcwidth = "0.2.13"
django-postgres-extra = "^2.0.8"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
